Line breaks added, replaced some code
Added line breaks to the end and start of some prints for less space wasted. Put elif's for vowels into one if statement. Changed ending question so it is chosen from a list. Changed the code for finding their answer, which also made the code easier to read. Trying to add code so the computer doesn't ask "What's an object that starts with  ?", still need to fix it.